Add: modules/credit.js
