+++ src/lottery/services/globalState.js
@@ -5,19 +5,6 @@ const {
 } = require("../modules/stores/mongo");
 const logger = require("../../modules/logger");
 
-const getUserCreditAmount = async (req) => {
-  const eventStatus = await eventStatusModel.findOne({ userId: req.userOid });
-  if (!eventStatus) return null;
-
-  return {
-    creditAmount: eventStatus.creditAmount,
-    creditUpdate: async (delta) => {
-      eventStatus.creditAmount += delta;
-      await eventStatus.save();
-    },
-  };
-};
-
 const getUserGlobalStateHandler = async (req, res) => {
   try {
     let eventStatus = await eventStatusModel.findOne({ userId: req.userOid });
